+++ android/src/main/java/io/openim/flutter_openim_sdk/FlutterOpenimSdkPlugin.java
@@ -18,6 +18,7 @@ import io.openim.flutter_openim_sdk.manager.FriendshipManager;
 import io.openim.flutter_openim_sdk.manager.GroupManager;
 import io.openim.flutter_openim_sdk.manager.IMManager;
 import io.openim.flutter_openim_sdk.manager.MessageManager;
+import io.openim.flutter_openim_sdk.manager.SignalingManager;
 import io.openim.flutter_openim_sdk.manager.UserManager;
 
 
@@ -37,6 +38,7 @@ public class FlutterOpenimSdkPlugin implements FlutterPlugin, MethodCallHandler 
     private static MessageManager messageManager;
     private static ConversationManager conversationManager;
     private static GroupManager groupManager;
+    private static SignalingManager signalingManager;
 
     public FlutterOpenimSdkPlugin() {
     }
@@ -48,6 +50,7 @@ public class FlutterOpenimSdkPlugin implements FlutterPlugin, MethodCallHandler 
         FlutterOpenimSdkPlugin.messageManager = new MessageManager();
         FlutterOpenimSdkPlugin.conversationManager = new ConversationManager();
         FlutterOpenimSdkPlugin.groupManager = new GroupManager();
+        FlutterOpenimSdkPlugin.signalingManager = new SignalingManager();
     }
 
     @Override