+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/FriendshipManager.java
@@ -112,16 +112,4 @@ public class FriendshipManager extends BaseManager {
                 jsonValue(methodCall)
         );
     }
-
-//    public void forceSyncFriendApplication(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.forceSyncFriendApplication();
-//    }
-//
-//    public void forceSyncFriend(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.forceSyncFriend();
-//    }
-//
-//    public void forceSyncBlackList(MethodCall methodCall, MethodChannel.Result result) {
-//        Open_im_sdk.forceSyncBlackList();
-//    }
 }